Add MCA file reading code
Add the code to actually process an MCA file (rail timetable info in
the Network Rail CIF format). It is based around a state machine
that attempts to ensure the file meets the requirements in the
available documentation. However as the files freely available from
data.atoc.org only use a subset of features, it will be hard to fully
test it.