Add function for creating (prepared) SQL insert statements
The code for creating the SQL insert statements in mca_reader is
repetitive so it is now a method in nrcif. The insert is now a
prepared statement in Postgresql, although the benefit of this is
likely to be minimal as I can't imagine the SQL parse time is
significant.