Add support for MSN station name files
Add support for the MSN master station name file as defined by
ATOC. This is built around the basic CIF support already provided.
Any record types marked as historic in the documentation have
been ignored and are not supported at all. Note that the header
(first line) has to be read and discarded before you can start
feeding records to the processing routine, as the header has the
same type indicator as a station detail record.