Add a flag to indicate that a train schedule has crossed midnight
Train schedules are always associated with the day that the train
starts, even if this is just before midnight and most of the stops
of the train happen on the following day. To track this a flag
xmidnight has been added to indicate that the times specified are
now for the following day. I don't think that there are likely to be
any trains that take more than 24 hours so this should be
sufficient.