Change back to using the ATOC location data for isochron
Using the NaPTAN data as a cross-reference, the datum for the
ATOC data can be found. After the offset, the differences in station
locations between the two data sets are between zero and
hundred metres in almost all cases. For drawing isochron maps
this is not significant, so by default the isochron function will not
require NaPTAN data to be loaded.