Add conversion between E, N and lat, lon co-ordinates
Add conversion function between Eastings and Northings and
lattitude and longitude pairs. The function provided is based on
UK OS National Grid and so the lattitudes and longitudes
produced are based on the Airey 1830 ellipsoid, not the GRS80
one used for GPS. There will therefore be residual differences if
these lat/lon pairs are plotted using a GRS80 based map.