Fix bug in util.isocron and reformat
The date passed into util.isocron would be used to create the correct
day's timetable, but in the call to iterate_reachable the date
field was filled out with a constant. This had a limited impact
because only the link with the ALF tables was affected. This has been
fixed and the file has also been reformatted.